remove some misleading comments, put SteensgaardAnalysis in its own file (should be deprecated?)
